App/stock_analysis_project/requirements.txt
alabaster==0.7.12
amqp==2.6.0
asgiref==3.2.10
autopep8==1.5.3
Babel==2.8.0
backcall==0.2.0
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.1rc1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-extensions==3.0.2
django-import-export==2.2.0
django-timezone-field==4.0
django-webpack-loader==0.7.0
djangorestframework==3.11.0
docutils==0.16
et-xmlfile==1.0.1
flower==0.9.5
future==0.18.2
humanize==2.5.0
idna==2.10
imagesize==1.2.0
ipython==7.16.1
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.1
Jinja2==2.11.2
kombu==4.6.11
lxml==4.5.2
MarkupPy==1.14
MarkupSafe==1.1.1
multitasking==0.0.9
mysqlclient==2.0.1
numpy==1.19.0
odfpy==1.4.1
openpyxl==3.0.4
packaging==20.4
pandas==1.0.5
parso==0.7.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
pycodestyle==2.6.0
Pygments==2.6.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
yfinance==0.1.54
